+++ src/d3cap/util.rs
@@ -8,6 +8,6 @@ pub unsafe fn skip_cast<T, U>(t: &T) -> &U {
     &*((t as *const T).offset(1) as *const U)
 }
 
-pub unsafe fn skip_bytes_cast<T,U>(t: &T, bytes: int) -> &U {
+pub unsafe fn skip_bytes_cast<T,U>(t: &T, bytes: isize) -> &U {
     &*((t as *const T as *const u8).offset(bytes) as *const U)
 }
